backend: don't fail rebuild-netboot when no kernel its defined
We also have qemu-based machines now, that won't need FIT images
to boot in all cases.

So don't fail when there is no kernel its defined and just emit a
message stating the fact.

+++ server/rebuild-netboot.sh
@@ -20,6 +20,11 @@ build_uimage_netboot() {
 build_netboot() {
 	local INST=$2
 
+	if [ ! -f /home/devel/nfs/kernel/$INST-kernel.its ]; then
+		echo "$INST: no netboot image to build"
+		return
+	fi
+
 	echo "$INST: building netboot image"
 	mkimage -D "-q" -f /home/devel/nfs/kernel/$INST-kernel.its /home/devel/nfs/kernel/tmp/$INST-vmlinux.uimg > /dev/null
 	cp /home/devel/nfs/kernel/tmp/$INST-vmlinux.uimg /home/devel/tftp/hstuebner/$INST.vmlinuz
